requirements.txt
beautifulsoup4>=4.13.0b2
certifi>=2024.8.30
charset-normalizer>=3.4.0
coloredlogs>=15.0.1
ephem>=4.1.6
humanfriendly>=10.0
idna>=3.10
lxml>=5.3.0
numpy==2.3.5
pillow>=9.5.0
pydub>=0.25.1
pyephem>=9.99
pygame>=2.6.1
pyserial>=3.5
python-dotenv>=0.21.1
pytz>=2024.2
requests>=2.31.0
curl_cffi>=0.13.0
six>=1.16.0
soupsieve>=2.4.1
tqdm>=4.66.6
urllib3>=2.0.7
pause>=0.3
pycryptodome>=3.23.0
